# swagger.yaml
openapi: 3.0.0
info:
  title: ISDServices
  description: In order to create a post on your medical reserch work, read about other medical post and comments, ISDServices provides you this placform, you must first initiate the process from the platform by registering.
  contact:
    name: hr.admin@isdservices.org
    email: integratedsoftwaredataservices@gmail.com
  version: "1.0"
servers:
  - url: https://isdservices.herokuapp.com
    description: This is the production server
    variables: {}
  - url: http://localhost:4400
    description: "This is the localhost development server"
    variables: {}
paths:
...
  /isds-admin/create-contact:
    post:
      tags:
        - User
      summary: contactUs
      description: contactUs
      operationId: contactUs
      parameters: []
      requestBody:
        description: "Send a message to ISDS admin"
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/contactUs"
                - example:
                    email: test@domain.com
                    message: put your message here
                    fullName: "Felix Adegbesan"
            example:
              email: test@domain.com
              message: put your message here
              fullName: "Felix Adegbesan"
        required: true
      responses:
        "200":
          description: ""
          headers: {}
          content: {}
      deprecated: false
      security: []
...
  /user/single-user:
    get:
      tags:
        - User
      summary: SingleUser
      description: This returns details of a single user
      operationId: SingleUser

      responses:
        "200":
          description: ""
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearer: []
...
  /user/register:
    post:
      tags:
        - Auth
      summary: RegisterUser
      description: This is the signup endpoint, which allows you to register a user
      operationId: RegisterUser
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/registeruserrequest"
                - example:
                    name: John
                    occupation: Doe
                    dateOfBirth: 04/01/2003
                    email: example@domain.com
                    password: "1234567890"
                    confirmPassword: "1234567890"
            example:
              name: John
              occupation: Doyin
              dateOfBirth: 04/01/2004
              email: example@domain.com
              password: "1234567890"
              confirmPassword: "1234567890"
        required: true
      responses:
        "200":
          description: ""
          headers: {}
          content: {}
      deprecated: false
      security: []
  /user/login:
    post:
      tags:
        - Auth
      summary: LoginUser
      description: You can log in either by using a registered email or a registered dateOfBirth with a password
      operationId: LoginUser
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/loginuserrequest"
                - example:
                    email: email@hostname.com
                    password: userpassword
            example:
              email: email@hostname.com
              password: userpassword
        required: true
      responses:
        "200":
          description: ""
          headers: {}
          content: {}
      deprecated: false
      security: []
  /user/verify/{token}:
    get:
      tags:
        - Auth
      summary: VerifyEmail
      description: This endpoint verifies an unverified email, to get the token for verification you can go through the resend verification token endpoint or upon registration of the user, after which the user would be redirected to the login page
      operationId: VerifyEmail
      parameters:
        - name: token
          in: path
          description: The user ID
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
          content: {}
      deprecated: false
      security: []
...
  /user/all-users:
    get:
      tags:
        - Admin
      summary: AllUsers
      description: This endpoint returns all registered users, accessed by logged-in admin only
      operationId: AllUsers
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearer: []
...
  /user/delete-user/{userId}:
    delete:
      tags:
        - Admin
      summary: DeleteUser
      description: This endpoint deletes a registered user from the system database, accessed only by logged-in admin
      operationId: DeleteUser
      parameters:
        - name: userId
          in: path
          description: The user ID
          required: true
          style: simple
          schema:
            type: string
      responses:
        "200":
          description: ""
          headers: {}
          content: {}
      deprecated: false
      security:
        - bearer: []
components:
  schemas:
...
    registeruserrequest:
      title: registeruserrequest
      required:
        - name
        - occupation
        - dateOfBirth
        - email
        - password
        - confirmPassword
      type: object
      properties:
        name:
          type: string
        occupation:
          type: string
        dateOfBirth:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
      example:
        name: Adeyinka
        occupation: Doctor
        dateOfBirth: 04/04/2023
        email: example@domain.com
        password: "1234567890"
        confirmPassword: "1234567890"
    loginuserrequest:
      title: loginuserrequest
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: example@domain.com
        password: "1234567890"
...
    contactUs:
      title: contactUs
      required:
        - email
        - message
        - fullName

      type: object
      properties:
        email:
          type: string
        message:
          type: string
        fullName:
          type: string

      example:
        email: test@domain.com
        userEmail: "test@des.pol"
        fullName: "this is good"
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...